+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Send complete HTML document in redirect & error responses
+  * Set default CSP header in redirect & error responses
 
 0.14.2 / 2017-01-23
 ===================